loaded BVH is good but the numbers are not consistent
Note: the GLB exported is expected to be T-Pose, instead of folded
when all of the quat channels are Euler(0,0,0).

But it's folded at a weired angle, so we need to multiple the rot
in by the reference rotation at T-Pose.
I believe there's something wrong on this step, I don't really know what
that is